+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+### Version 20260305
+- Add per-page request count metrics
+
 ### Version 20260304
 - Added graceful shutdown support: All editing sessions will be draft-saved before the server shuts down.
 - Added local save on clients. Saves will be stored on browser local storage, and will be automatically loaded when users return to the editor. This provides a safety net against accidental data loss due to network issues or server problems.